+++ Backend/app/dashboard/rider-dashboard/src/API/RiderPlatform/RideBooking/Booking.hs
@@ -16,7 +16,7 @@ module API.RiderPlatform.RideBooking.Booking where
 
 import qualified "rider-app" API.Dashboard.RideBooking.Booking as BAP
 import qualified Dashboard.RiderPlatform.Customer as Common
-import qualified "rider-app" Domain.Action.UI.Booking as DBooking
+import qualified "rider-app" Domain.Types.Booking as DBooking
 import qualified "rider-app" Domain.Types.Booking as SRB
 import qualified "rider-app" Domain.Types.Booking.API as DB
 import qualified "lib-dashboard" Domain.Types.Merchant as DM
