+++ Backend/app/rider-platform/rider-app/Main/src/API/UI/Booking.hs
@@ -13,7 +13,7 @@
 -}
 
 module API.UI.Booking
-  ( DBooking.BookingListRes,
+  ( BookingListRes,
     API,
     handler,
     bookingStatus,
@@ -22,7 +22,7 @@ module API.UI.Booking
 where
 
 import qualified Domain.Action.UI.Booking as DBooking
-import Domain.Types.Booking (BookingAPIEntity)
+import Domain.Types.Booking
 import qualified Domain.Types.Booking as SRB
 import qualified Domain.Types.Person as Person
 import Environment
@@ -43,7 +43,7 @@ type API =
              :> QueryParam "offset" Integer
              :> QueryParam "onlyActive" Bool
              :> QueryParam "status" SRB.BookingStatus
-             :> Get '[JSON] DBooking.BookingListRes
+             :> Get '[JSON] BookingListRes
        )
 
 handler :: FlowServer API
@@ -54,5 +54,5 @@ handler =
 bookingStatus :: Id SRB.Booking -> Id Person.Person -> FlowHandler BookingAPIEntity
 bookingStatus bookingId = withFlowHandlerAPI . DBooking.bookingStatus bookingId
 
-bookingList :: Id Person.Person -> Maybe Integer -> Maybe Integer -> Maybe Bool -> Maybe SRB.BookingStatus -> FlowHandler DBooking.BookingListRes
+bookingList :: Id Person.Person -> Maybe Integer -> Maybe Integer -> Maybe Bool -> Maybe SRB.BookingStatus -> FlowHandler BookingListRes
 bookingList personId mbLimit mbOffset mbOnlyActive = withFlowHandlerAPI . DBooking.bookingList personId mbLimit mbOffset mbOnlyActive
