+++ Backend/app/rider-platform/rider-app/Main/src/API/UI/Confirm.hs
@@ -23,6 +23,7 @@ where
 
 import qualified Beckn.ACL.Init as ACL
 import qualified Domain.Action.UI.Confirm as DConfirm
+import Domain.Action.UI.SimulatedFlow.Confirm
 import qualified Domain.Types.Person as SP
 import qualified Domain.Types.Quote as Quote
 import Environment
@@ -33,7 +34,6 @@ import Kernel.Utils.Common
 import Kernel.Utils.Error.BaseError.HTTPError.BecknAPIError
 import Servant
 import qualified SharedLogic.CallBPP as CallBPP
-import SharedLogic.SimulatedFlow.Confirm
 import qualified Storage.Queries.Person as QP
 import Tools.Auth
