+++ Backend/app/rider-platform/rider-app/Main/src/Domain/Action/UI/Maps.hs
@@ -27,6 +27,7 @@ where
 
 import qualified Data.Geohash as DG
 import Data.Text (pack)
+import qualified Domain.Action.UI.SimulatedFlow.Maps as SF
 import Domain.Types.Maps.PlaceNameCache as DTM
 import qualified Domain.Types.Merchant as DMerchant
 import qualified Domain.Types.Person as DP
@@ -37,7 +38,6 @@ import Kernel.Storage.Esqueleto as Esq
 import Kernel.Types.Error
 import Kernel.Types.Id
 import Kernel.Utils.Common
-import qualified SharedLogic.SimulatedFlow.Maps as SF
 import qualified Storage.CachedQueries.CacheConfig as SCC
 import qualified Storage.CachedQueries.Maps.PlaceNameCache as CM
 import qualified Storage.CachedQueries.Merchant as QMerchant
@@ -48,9 +48,7 @@ import Tools.Metrics (CoreMetrics)
 autoComplete :: (EncFlow m r, EsqDBFlow m r, SCC.CacheFlow m r, CoreMetrics m) => Id DP.Person -> Maps.AutoCompleteReq -> m Maps.AutoCompleteResp
 autoComplete personId req = do
   person <- QP.findById personId >>= fromMaybeM (PersonNotFound personId.getId)
-  if person.isSimulated
-    then SF.autoComplete req
-    else Maps.autoComplete person.merchantId req
+  SF.autoCompleteSimulator person.isSimulated req $ Maps.autoComplete person.merchantId req
 
 getPlaceDetails :: (EncFlow m r, EsqDBFlow m r, SCC.CacheFlow m r, CoreMetrics m) => Id DP.Person -> Maps.GetPlaceDetailsReq -> m Maps.GetPlaceDetailsResp
 getPlaceDetails personId req = do
