+++ Backend/app/rider-platform/rider-app/Main/src/Domain/Action/UI/Ride.hs
@@ -20,6 +20,7 @@ module Domain.Action.UI.Ride
   )
 where
 
+import Domain.Action.UI.SimulatedFlow.Ride
 import Domain.Types.Booking.BookingLocation (BookingLocationAPIEntity, makeBookingLocationAPIEntity)
 import qualified Domain.Types.Booking.Type as DB
 import qualified Domain.Types.Person as SPerson
@@ -36,8 +37,6 @@ import Kernel.Utils.CalculateDistance (distanceBetweenInMeters)
 import Kernel.Utils.Common
 import qualified SharedLogic.CallBPP as CallBPP
 import Storage.CachedQueries.CacheConfig
-import qualified Storage.CachedQueries.SimulatedFlow.Driver as CD
-import qualified Storage.CachedQueries.SimulatedFlow.SearchRequest as CSSearchRequest
 import qualified Storage.Queries.Booking as QRB
 import qualified Storage.Queries.Person as QP
 import qualified Storage.Queries.Person as QPerson
@@ -46,8 +45,6 @@ import Tools.Error
 import qualified Tools.Maps as MapSearch
 import Tools.Metrics
 import qualified Tools.Notifications as Notify
-
-type GetDriverLocResp = MapSearch.LatLong
 
 data GetRideStatusResp = GetRideStatusResp
   { fromLocation :: BookingLocationAPIEntity,
@@ -74,56 +71,7 @@ getDriverLocWrapper ::
   m GetDriverLocResp
 getDriverLocWrapper rideId personId = do
   person <- runInReplica $ QPerson.findById personId >>= fromMaybeM (PersonNotFound personId.getId)
-  if not person.isSimulated
-    then getDriverLoc rideId personId
-    else do
-      bookingId <- CSSearchRequest.getBookingIdByRideId rideId >>= fromMaybeM (InvalidRequest "SimulatedFlow:Booking: booking not found")
-      quoteId <- CSSearchRequest.getQuoteIdByBookingId bookingId >>= fromMaybeM (InvalidRequest "SimulatedFlow:Quote: quote not found")
-      estimateId <- CSSearchRequest.getEstimateIdByQuoteId quoteId >>= fromMaybeM (InvalidRequest "SimulatedFlow:Estimate: estimate not found")
-      selectedDriver <- CD.getLinkedDriverByEstimateId estimateId >>= fromMaybeM (InvalidRequest "SimulatedFlow:RouteInfo: route not found")
-      locationInfo <- CSSearchRequest.getSimulatedLocationInfoByRideId rideId
-      now <- getCurrentTime
-      let driverRoute = selectedDriver.driverRoute
-      case locationInfo of
-        Nothing -> do
-          let newLocationInfo =
-                SimulatedLocationInfo
-                  { lastCacheTimeStamp = now,
-                    indexOfRouteInfo = 0,
-                    doneCount = 0
-                  }
-          CSSearchRequest.cacheSimulatedLocationInfoByRideId rideId newLocationInfo
-          return $ head driverRoute.points
-        Just info -> do
-          let routeDuration = fromMaybe 60 driverRoute.duration
-          let indDiff = round (diffUTCTime now info.lastCacheTimeStamp) `div` max 1 routeDuration.getSeconds
-              newIndex = indDiff + info.indexOfRouteInfo
-          if newIndex < length driverRoute.points
-            then do
-              let newLocationInfo =
-                    SimulatedLocationInfo
-                      { lastCacheTimeStamp = now,
-                        indexOfRouteInfo = newIndex,
-                        doneCount = info.doneCount
-                      }
-              CSSearchRequest.cacheSimulatedLocationInfoByRideId rideId newLocationInfo
-              return $ driverRoute.points !! newIndex
-            else do
-              let doneCount = info.doneCount + 1
-              let newLocationInfo =
-                    info
-                      { lastCacheTimeStamp = now,
-                        doneCount = doneCount
-                      }
-              thresholdCount <- asks (.simulatedMaxDone)
-              if doneCount > thresholdCount
-                then do
-                  CSSearchRequest.linkBookingStatusBooking DB.CANCELLED bookingId
-                  CSSearchRequest.cacheSimulatedLocationInfoByRideId rideId newLocationInfo
-                  return $ last driverRoute.points
-                else do
-                  CSSearchRequest.cacheSimulatedLocationInfoByRideId rideId newLocationInfo
-                  return $ last driverRoute.points
+  simulateRide person.isSimulated rideId $ getDriverLoc rideId personId
 
 getDriverLoc ::
   ( HasCacheConfig r,
