+++ Backend/app/rider-platform/rider-app/Main/src/Domain/Action/UI/Route.hs
@@ -21,6 +21,7 @@ module Domain.Action.UI.Route
   )
 where
 
+import Domain.Action.UI.SimulatedFlow.Route
 import qualified Domain.Types.Person as DP
 import Kernel.Prelude
 import Kernel.Types.Error (PersonError (PersonNotFound))
@@ -34,20 +35,14 @@ import Tools.Metrics (CoreMetrics)
 getRoutes :: (EncFlow m r, CacheFlow m r, EsqDBFlow m r, CoreMetrics m) => Id DP.Person -> Maps.GetRoutesReq -> m Maps.GetRoutesResp
 getRoutes personId req = do
   person <- QP.findById personId >>= fromMaybeM (PersonNotFound personId.getId)
-  if person.isSimulated
-    then Maps.getSimulatedRoutes person.merchantId req
-    else Maps.getRoutes person.merchantId req
+  simulateRoute person.isSimulated person.merchantId req $ Maps.getRoutes person.merchantId req
 
 getPickupRoutes :: (EncFlow m r, CacheFlow m r, EsqDBFlow m r, CoreMetrics m) => Id DP.Person -> Maps.GetRoutesReq -> m Maps.GetRoutesResp
 getPickupRoutes personId req = do
   person <- QP.findById personId >>= fromMaybeM (PersonNotFound personId.getId)
-  if person.isSimulated
-    then Maps.getSimulatedRoutes person.merchantId req
-    else Maps.getPickupRoutes person.merchantId req
+  simulateRoute person.isSimulated person.merchantId req $ Maps.getRoutes person.merchantId req
 
 getTripRoutes :: (EncFlow m r, CacheFlow m r, EsqDBFlow m r, CoreMetrics m) => Id DP.Person -> Maps.GetRoutesReq -> m Maps.GetRoutesResp
 getTripRoutes personId req = do
   person <- QP.findById personId >>= fromMaybeM (PersonNotFound personId.getId)
-  if person.isSimulated
-    then Maps.getSimulatedRoutes person.merchantId req
-    else Maps.getTripRoutes person.merchantId req
+  simulateRoute person.isSimulated person.merchantId req $ Maps.getRoutes person.merchantId req
